Add gitlab CI configuration file (!15)
This CI config is based on the travis one. nsiqcppstyle was dropped as
it required python2.

In its place, two clang-based tools were introduced: clang-tidy for some
static analysis, and clang-format for some linting.

Also delete the travis configuration, as it is stale, unlikely to be
used anymore, and prevent it from rotting.